# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API Sistem Manajemen Parkir.",
        "title": "Parking API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
      "/vehicles": {
        "get": {
          "description": "Get list of vehicles with pagination",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Vehicles"
          ],
          "summary": "Get All Vehicles",
          "parameters": [
            {
              "type": "integer",
              "default": 1,
              "description": "Page number",
              "name": "page",
              "in": "query"
            },
            {
              "type": "integer",
              "default": 10,
              "description": "Items per page",
              "name": "limit",
              "in": "query"
            },
            {
              "type": "string",
              "description": "Search by plate_number or description",
              "name": "search",
              "in": "query"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.PageResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "post": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Vehicles"
          ],
          "summary": "Create New Vehicle",
          "parameters": [
            {
              "description": "Vehicle Request",
              "name": "request",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/requests.CreateVehicleRequest"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
      "/vehicles/{id}": {
        "get": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Vehicles"
          ],
          "summary": "Get Vehicle By ID",
          "parameters": [
            {
              "type": "string",
              "description": "Vehicle Hash ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "put": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Vehicles"
          ],
          "summary": "Update Vehicle",
          "parameters": [
            {
              "type": "string",
              "description": "Vehicle Hash ID",
              "name": "id",
              "in": "path",
              "required": true
            },
            {
              "description": "Update Request Body",
              "name": "request",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/requests.UpdateVehicleRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "delete": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Vehicles"
          ],
          "summary": "Delete Vehicle",
          "parameters": [
            {
              "type": "string",
              "description": "Vehicle Hash ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "requests.CreateVehicleRequest": {
        "type": "object",
        "required": [
          "plate_number",
          "vehicle_type_id"
        ],
        "properties": {
          "customer_id": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "maxLength": 1000
          },
          "plate_number": {
            "type": "string",
            "maxLength": 20
          },
          "vehicle_type_id": {
            "type": "string"
          }
        }
      },
...
      "requests.UpdateVehicleRequest": {
        "type": "object",
        "properties": {
          "customer_id": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "maxLength": 1000
          },
          "plate_number": {
            "type": "string",
            "maxLength": 20
          },
          "vehicle_type_id": {
            "type": "string"
                }
            }
        },
...
        "responses.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "responses.Meta": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "current_url": {
                    "description": "Navigasi URL",
                    "type": "string"
                },
                "filters": {
                    "type": "object",
                    "additionalProperties": true
                },
                "first_page": {
                    "description": "Navigasi Angka",
                    "type": "integer"
                },
                "first_url": {
                    "type": "string"
                },
                "last_page": {
                    "type": "integer"
                },
                "last_url": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "next_page": {
                    "type": "integer"
                },
                "next_url": {
                    "type": "string"
                },
                "prev_page": {
                    "description": "Pakai pointer biar bisa null kalau gak ada",
                    "type": "integer"
                },
                "prev_url": {
                    "type": "string"
                },
                "search": {
                    "type": "string"
                },
                "sort": {
                    "description": "Search \u0026 Filter",
                    "type": "string"
                },
                "total_data": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "responses.PageResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/responses.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}